// src/app/shared/models/router-history.ts
// CREDIT: https://github.com/semanticbits/previous-url-example
import { HistoryEntry } from './history-event';
import { NavigationTrigger } from './navigation-trigger';

import { NavigationStart, NavigationEnd } from '@angular/router';

export interface RouterHistory {
  history: HistoryEntry[];
  currentIndex: number;

  event: NavigationStart | NavigationEnd;
  trigger: NavigationTrigger;
  id: number;
  idToRestore: number;
}